+++ tests/conftest.py
@@ -64,7 +64,8 @@ def db_engine():
         ApplicationTemplateQuestion,
     )
     from models.reminder import ReminderMessage  # noqa: F401
-    from models.admin import BotModeratorRole, PermissionSubscriber, GuildLanguageConfig  # noqa: F401
+    from models.guild import GuildLanguageConfig  # noqa: F401
+    from models.permissions import BotModeratorRole, PermissionSubscriber  # noqa: F401
     from models.leavemsg import LeaveMessage  # noqa: F401
     from models.moderation import AutoDelete, AutoKicker  # noqa: F401
     from models.reactionrole import ReactionRoleEntry, ReactionRoleMessage  # noqa: F401
